schematrons: Check for empty elements in rule context (TEDEFO-419)
Make the rule context match elements that have no child and are empty,
and set the test as "false".
This is equivalent to what was done before, but avoids having a lot of
pointless and identical "fired-rule" elements in the validation reports.

Also try to make the message a bit easier to read.