schematrons: Use label codes in all generated rules
The English text in the assert element is replaced with the code of the
corresponding label (TEDEFO-290).

Translations of all these labels in all EU official languages will be
available in the "translations" folder, in the "rule_*.xml" files.

This also changes some values of the "id" attribute in stage-4.sch.

Apart from this, this commit does not change the rules themselves.