schematrons: New handling of "not repeatable" and "multilingual" rules
The notions of "not repeatable" and "multilingual" are now considered as
characteristics of the field. So in stage-2-repeatable, the assert id is
not a rule identifier anymore, but is now based on the field id.

This does not changes the rules themselves, but some of them are in a
different place in the file, as they are ordered on the id.