schematrons: Rework change notice rules
Remove unnecessary checks on notice subtype: if the field must be
unchanged whenever it is present, no need to exclude notice subtype
where the field is not allowed. This is enforced by other rules.

Don't check separately for items added and removed, do it in a single
assert.

Remove unneeded parentheses in test expressions.

Reformat the file for readability.