schematrons: Simplify checks on set of values for change notices
Instead of checking that the number of values is the same, and then
checking that each value from the change notice is in the parent notice,
we just do a "deep equal" on the sorted sequences.

With this approach there is a single assert, instead of 2 asserts in 2
different contexts.

Also remove some rules that are now obsolete or not useful for now, as
we're not trying to be exhaustive for the moment.