schematrons: Fix test for added fields in change notices
If a field that must not be changed is not present in the parent notice
but is present in the change notice, this should cause a validation
failure. But it was not the case.

Fix this by using "deep-equal", as it behaves as we need:
* If A and B do not exist, deep-equal(A, B) is true
* If one exists and the other not, deep-equal(A, B) is false
This seems a bit overkill, but it's nicer than complicating further the
XPath expressions.