schematrons: Add diagnostic information in all generated rules
The location of a failure is indicated in validation reports based on
the context on the rule. So for rules that have a context that does not
point to a leaf element (a field), we add the XPath of the leaf element,
relative to the context of the rule.

This information is added with the "diagnostics" feature in Schematron:
- All diagnostic information items are defined by an id and a text, and
they are centralized at the top-level of the Schematron.
- Each assert can reference one or more of those items in the
"diagnostics" attribute.

With this change, the file "entry.sch" is now also generated, to add the
diagnostic information automatically.

Update affected reports, to add the "diagnostic-reference" element in
all failed asserts.