schematrons: Remove stage-4-custom.sch and update stage 5
The rules that were in stage-4-custom.sch are now covered by the
rules in the database, in particular with the ones added in stage 5 in
this commit.

This coverage is not fully complete, there are a few rules that are not
generated yet, due to technical issues. But they will be added soon.

Update affected validation reports, no real change.