fields: Update node structure (TEDEFO-1463)
Add several nodes, and change the parent node of fields, so that the
structure better corresponds to the "conceptual" or visual structure of
notices.

For affected fields, the "xpathRelative" property is adapted, and the
"repeatable" property when needed. As field are still in the same place
in the XML, this does not change their "xpathAbsolute", except for a few
where a predicate is added or moved.