fields: Add 2 missing codes and codelist constraints (TEDEFO-1658)
The fields BT-195(BT-635)-Lot and BT-195(BT-636)-Lot have type "code"
but the codelist constraints were missing, and the needed codes were
missing from the "non-publication-identifier" codelist.

Add the 2 codes, the 2 constraints, the corresponding schematron
rules, and the texts in EN for the codes and rules.

Also update affected validation report.