fields: Fix uniqueness assert on BT-150-Contract (TEDEFO-1607)
Checking that the value in BT-150-Contract is unique should only be done
when the field is present. Change the rule to have the field itself
as context, so it is not fired when the field is not present.

Cherry picked from commit 298d4425, to solve this problem in 1.3.x, as
discussed in #334.