fields: Correct a constraint for notice subtype 28
The assert "BR-BT-00137-0209" on BT-137 would fail when
BT-13713-LotResult is absent.
As BT-13713-LotResult is forbidden for notice subtype 28, this made it
impossible to create a valid notice with subtype 28.

Fix the condition for this assert so that it is not applied when
BT-13713-LotResult is not present.

Also apply the corresponding change in the Schematron rules.

Cherry picked from commits 1c48ad27 and 6b09cc586 that are part of
version 1.4.0, and edited to keep only the minimal change to fix the
issue described in #316