fields: Add new field OPT-060-Lot for terms performance (TEDEFO-1925)
BT-70-Lot must have the following as a sibling in the XML:
<cbc:ExecutionRequirementCode listName="conditions">performance</cbc:ExecutionRequirementCode>

So add a new technical field OPT-060-Lot to specify this, with a preset
value. Also add the corresponding rules. This includes new conditional
rules, to ensure that BT-70-Lot and OPT-060-Lot are present together.

Update comments in cn_24_maximal.xml to indicated this new field.

Update affected reports, with just changes in fired rules.