codelists: Add parent code in nuts codelist (TEDEFO-1755)
Add a new column "parentCode" in teh main Nuts codelist. For each code
that is not at the top of the hierarchy, gives the identifier of the
parent code.

This add codes for countries: although they are not in the NUTS
classification, they are the roots of the structure, so they are useful
when displaying the classification in a tree-like view.

Also update corresponding tailored codelists based on NUTS.