examples: Use only static rules for the validation reports
The newly added dynamic rule that checks the value of the dispatch date
will fail for all the examples, because their dispatch date is fixed.
So we know only use the static rules to get the validation reports.
This removes some relevant part in the reports for change notices, so
we'll see if we can bring those back.