schematrons: Fix test expression in rule on dispatch date (TEDEFO-2232)
The change in the test expression in commit 735cc51a had an incorrect
syntax (using "<").
So go back to the previous expression, and just allow a dispatch date up
to 2 days in the past, to account for possible timezone differences.

This is more permissive, a more exact check will be done later.