fields: Allow non-ASCII characters in URL pattern (TEDEFO-2274)
We now use the same pattern as for the old TED schemas.

Please note this is done using the "\w" shorthand character class.
In XPath this matches all characters, not only ASCII
(XPath re-uses the spec for XSD:
https://www.w3.org/TR/xmlschema-2/#nt-MultiCharEsc)
But this might not be the case in other regex implementations.

Fixes issue #479.