fields: Revert BT-702(b)-notice to previous definition (TEDEFO-2431)
Having the field BT-702(b)-notice under a repeatable node is not useful,
and causes problems because it is part of the "metadata" in the notice
type definitions.

So we go back to its previous definition, with the field itself
repeatable.

Update the notice type definitions, to add back the "_repeatable"
property.

Also update the corresponding schematron rules, and the affected
validation report.

Fixes issue #549.