schematrons: Update manual rules for change notices (TEDEFO-2433)
The list of fields that cannot be changed in a change notice has been
reduced, so update the manual rules in validation-stage-6b.sch to not be
in contradiction.

As those manual rules are not the proper solution anyways, and are
currently not enforce by CVS, we only implement them for a few fields:
OPP-070-notice, BT-04-notice, BT-105-Procedure and BT-23-Procedure.

The other fields are more difficult to check fully correctly, so we
don't have them as manual rules and will wait for a proper solution.
It's better to have missing rules than to have rules that are incorrect
or inconsistent with what we really want to have.

This also solves the issue described in #544.