translations: Add texts in all languages for NUTS codes (TEDEFO-2591)
Until now, texts for NUTS codes were only provided in code_en.xml.
But those texts are not really English, they are the name of the region
in its local language (for code BG311 the label is "Видин").

We now provide those texts in all translations/code_*.xml files, so that
you don't have to have a special case to always look in code_en.xml for
NUTS codes.

Also add those labels in the  codelists/nuts*.gc, to be consistent.