schematrons: Restructure schematron files to use phases (TEDEFO-2660)
Split the schematron files further based on the notice subtype the rules
apply to, so that one pattern contains either rules that apply to one
specific notice subtype, or rules that are not specific to a notice
subtype. There is still one pattern per file.

The files "complete-validation.sch" defines one phase for each subtype.
Each phase references the corresponding patterns so that it has all the
rules that must be applied to notices of that subtype: all patterns for
the subtype, and all patterns not specific to a subtype.

When validating a notice, if you use the phase corresponding to the
subtype of the notice being validated, you will avoid executing the
XPaths for all the rules that apply only to other subtypes.
This reduces significantly the execution time for large notices, by a
factor of 6 in some examples.

This is backwards-compatible for existing usage of the Schematron:
the loading mechanism is the same, and if no phase is defined when
validating, all rules will be applied, as before.

Update all validation reports. This adds the indication of the phase
used, updates the pattern identifiers, and moves around fired-rule
elements. No real impact on the result of validation.