fields: Remove codelist constraints for BT-105-Procedure (TEDEFO-2857)
We added codelist constraints for BT-105-Procedure to indicate that only
some procedure types are allowed depending on the notice subtype.

As this is the first time we use constraints in the "codeList" property,
it can cause problem for some applications.

So we remove them from fields.json, going back to the previous
situation: BT-105-Procedure uses the "procurement-procedure-type"
codelist.

The schematron rules are left unchanged, so rules on procedure type and
notice subtype are still enforced during validation.