schematrons: Move a few elements to conform to ISO schema
The RELAX NG schema in the Schematron SIO standard requires some
elements to be in a certain order: title, ns, etc.
Move a few elments to avoid any errors when validating against this
schema.

This has no effect on the execution of the Schematron, implementations
seems less strict than the schema.