schematrons: Remove request to TED website (TEDEFO-2723)
Making a GET request to search for a notice is not possible with the
new TED website, as it supports only POST requests.

So remove the rule that makes this request, to avoid pointless requests
and silent errors.

(cherry picked from commit d14f55700ca8ddf7e13be289213c5bf129bf197e)