schematrons: Optimize CM rules for winner size (TEDEFO-3052)
The condition to determine if an organisation is a "winner" is very slow
on notices with a high number of tenders or contracts.
Move the corresponding schematron rules to validation-stage-4m.sch and
make them more efficient by using variables.

Update affected validation reports, adding fired rules in stage 4m.