schematrons: Optimize CF rules for BT-720-Tender (TEDEFO-3604)
The condition for when BT-720-Tender is forbidden can be slow for
notices with lots of tenders. Move the corresponding rules to
validation-stage-4m.sch and make them more efficient by using variable.

Although it was not generated yet, the rule for subtype E4 is included
in the manual rules, to avoid forgetting it later.