efx-grammar: Remove unneeded alternative in a parser rule (TEDEFO-3718)
In the fieldReference parser rule, having a direct reference to
fieldReferenceWithFieldContextOverride is not needed, because
fieldReferenceInOtherNotice can also match it with a chain of rules.

So removing this unneeded reference removes the ambiguity in the
rule, which reduces the parsing time, as the parser does not spend
time looking ahead and then backtracking.

This does not change the result of the translation.