schematrons: Fix error when one duplicate id is present
Due to a limitation in ph-schematron-pure, if an identifier is
duplicated only once in the XML notice, the corresponding
$global-...-duplicates variable would only contain one value, causing it
to not be evaluated as a sequence. This then causes an error in the
corresponding rule test.

Remove the use of distinct-values(...) to work around this. The
variable $global-...-duplicates will then contain the duplicate
identifiers repeated for each occurrence in the notice XML, so it will
always be a sequence.