workflows: Schedule benchmark to run every few days
The previous benchmark results are stored in the GitHub Cache, but cache
entries are deleted if they are not accessed in over 7 days.
And if the cache is empty, we will not detect that the rules are slower.

So schedule the benchmark to run automatically every Tuesday and Friday
at 05:00 in the morning, to avoid this deletion.  It might also be
helpful as a double-check.
The scheduled run will be on the develop branch.