schematrons: Remove incorrect stage 1 rules (TEDEFO-3873)
Stage 1 rules should only check the presence of XML elements that are
used as part of the context of a stage 2 rule (mandatory or forbidden
without condition).

This was not done correctly, causing some extra stage 1 rules to be
generated. Most of those extra rules were just redundant, except for one
rule on cac:SpecificTendererRequirement for subtype E3, due to a
specific situation (BT-71-Lot is optional for E3).
This rule caused an incorrect failure when BT-761-Lot was filled in.

Remove all incorrect/redundant rules.

Also update affected validation reports, to remove the corresponding
fired rules.