notice-types: Use labels with asset ids starting with GR- (TEDEFO-3830)
Instead of using node identifers when the group is associated with a
node, always use the group identifier as the label asset identifier.

Update the corresponding translations so that the texts corresponding to
each group are the same: the texts that were for "group|name|ND-..." are
copied over to the corresponding "group|name|GR-..." entry.
The "ND-... entries are removed, as they are not used anymore.

So if you use the NTDs and the translations, there will be no difference
in the texts shown in the UI.

Some "description" properties are modified in the NTDs, but those are
only for human reading the NTDs.