Publish via the new Maven Central Portal
Also move the definition of the repository for snapshot dependencies
from a separate "settings.xml" file into pom.xml.
This snapshot repository is only used to get snapshot versions of
eforms-sdk-analyzer, and defining it in a separate file does not have
any benefit, and might just cause it to be overlooked.
This also make the "analyze" and "benchmark" workflows more consistent
with the "publish" workflow, which did not use settings.xml.