schematrons: Remove redundant stage 1 rule on ND-FSR
Checking that cac:ContractExecutionRequirement is present is not needed,
as the rules on the fields under that element are executed in the
context of an element that is higher up in the hierarchy.

Correcting the logic that generates the stage 1 rules for this case also
changes a few of the node identifiers to which some of the rules
correspond. This has no real effect, as the new node corresponds to the
same XML element.

Fixes issue #1168.